requirements.txt
openai==1.88.0
gitpython==3.1.44
python-dotenv==1.0.0
typing-extensions==4.14.0